+++ nl/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix encoding of nested attributes and mark them with `NLA_F_NESTED`.
 - Support encoding and decoding of nested type-value attributes.
+- Support multiplexing requests via `Family.async`
 
 ## v0.2.4 (2026-08-22)